backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
google-genai>=0.1.0
pillow>=10.2.0
python-multipart>=0.0.9
pydantic>=2.6.0
python-dotenv>=1.0.1
